Fix a double free problem in the generated C code. (#300)
When the generated C code is doing an array initialization,
the initialization of one of the array members may fail.  If
that happens with the current code, then it will call "__destroy"
on each of the already-initialized members.  In turn, the
"__destroy" method will uninitialize each of the members of
the structure, then free the whole structure.  The final
freeing is incorrect; the freeing of the memory should be
done when the whole array is freed, not on a member-by-member
basis.  The solution here is to call the "__fini" method during
a failure, which will uninitialize members, but not free the
entire structure.  This should get rid of double-free warnings
that come from clang static analysis.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>