Update guard against common Windows preprocessor definitions
Rather than undefining unconditionally at the top of the file, only undefine
if a constant by one of the common names is defined.
Also, redefine the macro after the constant has been declared using the pragma
push_macro and pop_macro.
Suppress a potential warning for popping a macro that was not previously pushed.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>