Fix misuses of input iterators in BoundedVector
Input iterators are single pass, so calling std::distance(first, last)
will consume the input and leave [first,last) as an empty range.

This change makes assign(InputIterator, InputIterator) and
insert(const_iterator, InputIterator, InputIterator) dispatch to
overloaded functions that choose the right overload based on the
iterator category. For forward iterators the original implementation is
used. For input iterators assign will construct a new vector and if that
doesn't throw, use swap to replace the contents of *this. For insert
each value is inserted at the end and if the bound is not exceeded
std::rotate is used to shift them to the desired position, otherwise the
inserted elements are removed from the end before the function exits.

This retains the strong exception safety guarantee for both operations,
so that exceeding the bound does not alter the existing contents of the
vector.

Fixes #486

Signed-off-by: Jonathan Wakely <github@kayari.org>