Stop running string copy tests for multiple sizes
There isn't any interesting data here, and we aren't expecting the
performance to regress based on the size alone. I'm using an `Arg`
instead of hard-coding the size so that the test name reflects that
size, which should make it easier to compare results should we decide to
change it the size someday.

Signed-off-by: Scott K Logan <logans@cottsay.net>