Strip action service suffixes from C include prefix
This adds logic for the two missing services that are part of an action: `send_goal` and `get_result`.
For example, "fibonacci__send_goal.h" is not a valid header and we should instead use "fibonacci.h".

C++ and Python generators are not affected by this change since they combine all parts of an
action definition into a single file.

But generators like rosidl_generator_java create separate files for the service and message
definitions composing an action, and can take advantage of this common logic.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>