Fix zero-variadic-macro-arguments warning when using clang.
When building this package with clang, we get a lot of
warnings like the following:

must specify at least one argument for '...' parameter of variadic macro

That comes mostly from the TYPED_TEST_SUITE macros in gtest,
the signature of which looks like:

Prior to C++20, C++ technically did not allow macros with variadic
arguments to be called with nothing for the last variadic part.
This is allowed in C++20, but we aren't ready to transition to
it yet.

So just add in an unused third "empty" argument anytime we
use these variadic arguments.  This will quiet the warning on
clang and have no effect on operation.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>