Switch to using fastjsonschema for schema validation.
The main reason to do this is because jsonschema (what
we currently use) depends on referencing, which depends
on rpds-py, which has C libraries.  That means that when
using the python debug interpreter on Windows, we would
have to build a custom pip package for it to work.

In contrast, fastjsonschema is pure python and has no
dependencies.  It is also available on all of our default
platforms.  So switch to using it here so things work on
Windows debug.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>