Add name and data_type traits for actions
This implements the `rosidl_generator_traits::name<>()` and
`rosidl_generator_traits::data_type<>()` trait functions for generated
actions, similarly to how it's done for messages and services.

Unit tests are added to verify this behaviour, and also extend the prior
tests to cover the data_type trait for all interface types.

Signed-off-by: Nathan Wiebe Neufeldt <wn.nathan@gmail.com>